create  functional of city

+++ src/main/java/com/moskalev/repositories/TruckRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TruckRepository extends JpaRepository<Truck,Integer> {
+public interface TruckRepository extends JpaRepository<Truck,Long> {
 }
